Move class outside
